Struct rsgenetic::pheno::Fitness

pub struct Fitness {
    // some fields omitted
}

Fitness is a wrapper struct around a 64-bit floating point type, implementing equality. This is useful to be able to use functions for sorting and comparison, without requiring the population to be cloned.

Methods

impl Fitness

fn new(val: f64) -> Fitness

Construct a new Fitness object.

fn abs(&self) -> Fitness

Calculate the absolute value of some Fitness.