Module rsgenetic::sim

Contains implementations of Simulators, which can run genetic algorithms.

Modules

seq
Contains a sequential Simulator implementation.

Enums

FitnessType
Whether to maximize or to minimize the fitness value

SelectionType
The type of parent selection.

Traits

Builder
A Builder can create new Simulation instances.

Simulation
A Simulation is an execution of a genetic algorithm.

Type Definitions

NanoSecond
Used to make it more clear that run time is defined in nanoseconds.