Enum rsgenetic::sim::SelectionType

pub enum SelectionType {
    Maximize {
        count: usize,
    },
    Tournament {
        num: usize,
        count: usize,
    },
    Stochastic {
        count: usize,
    },
    Roulette {
        count: usize,
    },
}

The type of parent selection.

Variants

Maximize

Select only the count * 2 best performing parents in terms of fitness.

Fields

count

Should be larger than 0 and smaller than half the population size.

Tournament

Perform tournament selection with tournament size count, running num tournaments. This yields num * 2 parents.

Fields

num

Indicates the number of tournaments. Should be larger than 0 and smaller than the population size.

count

Should be larger than 0 and smaller than the population size.

Stochastic

Perform Stochastic Universal Sampling to do the selection. Selects count parents.

Fields

count

Should be larger than 0 and smaller than the population size.

Roulette

Perform Roulette Wheel Selection, also known as Fitness Proportionate Selection. This yields count parents.

Fields

count

Should be larger than 0 and smaller than the population size.