Struct rsgenetic::sim::select::MaximizeSelector

pub struct MaximizeSelector {
    // some fields omitted
}

Selects best performing phenotypes from the population.

Methods

impl MaximizeSelector

fn new(count: usize) -> MaximizeSelector

Create and return a maximizing selector.

Such a selector selects only the count best performing phenotypes as parents.

count: must be larger than zero, a multiple of two and less than the population size.