Struct rsgenetic::sim::select::StochasticSelector

pub struct StochasticSelector {
    // some fields omitted
}

Selects phenotypes at random, starting from a random index and taking equidistant jumps.

Commonly known as Stochastic Universal Sampling.

Methods

impl StochasticSelector

fn new(count: usize) -> StochasticSelector

Create and return a stochastic selector.

Such a selector selects elements using stochastic universal sampling, yielding parents with low, medium and high fitness values. In total, count parents are selected.

count: must be larger than zero, a multiple of 2 and less than the population size.