Struct rsgenetic::sim::select::TournamentSelector

pub struct TournamentSelector { /* fields omitted */ }

Runs several tournaments, and selects best performing phenotypes from each tournament.

Methods

impl TournamentSelector

fn new(count: usize, participants: usize) -> TournamentSelector

Create and return a tournament selector.

Such a selector runs count / 2 tournaments, each with participants participants. From each tournament, the best 2 phenotypes are selected, yielding count parents.

count: must be larger than zero, a multiple of two and less than the population size.

participants: must be larger than zero and less than the population size.