Trait rsgenetic::sim::select::Selector

pub trait Selector<T, F>: Debug where T: Phenotype<F>, F: Fitness {
    fn select(&self, population: &[T]) -> Result<Parents<T>, String>;
}

A Selector can select Parents for a new iteration of a Simulation.

Required Methods

fn select(&self, population: &[T]) -> Result<Parents<T>, String>

Select elements from a population for breeding.

If invalid parameters are supplied or the algorithm fails, this function returns an Err(String), containing a message indicating the error.

Otherwise it contains a vector of parent pairs wrapped in Ok.