Struct rsgenetic::sim::seq::SimulatorBuilder

pub struct SimulatorBuilder<'a, T, F> where T: 'a + Phenotype<F>, F: Fitness {
    // some fields omitted
}

A Builder for the Simulator type.

Methods

impl<'a, T, F> SimulatorBuilder<'a, T, F> where T: Phenotype<F>, F: Fitness

fn set_selector(self, sel: Box<Selector<T, F>>) -> Self

Set the selector of the resulting Simulator.

Returns itself for chaining purposes.

fn set_max_iters(self, i: u64) -> Self

Set the maximum number of iterations of the resulting Simulator.

The Simulator will stop running after this number of iterations.

Returns itself for chaining purposes.

fn set_early_stop(self, delta: F, n_iters: u64) -> Self

Set early stopping. If for n_iters iterations, the change in the highest fitness is smaller than delta, the simulator will stop running.

Returns itself for chaining purposes.