Module rsgenetic::sim

Contains implementations of Simulators, which can run genetic algorithms.

Modules

select
The selection module provides a trait that can be implemented to implement new selection algorithms. This module also provides a couple of useful selection algorithms.

seq
Contains a sequential implementation of ::sim::Simulation, called a Simulator.

types
This module provides implementations of the Fitness trait for some primitive types, such as i32, i64 etcetera. This is because Rust does not allow programmers to implement a foreign trait for a foreign type, which would stop you as a library user from using primitive types as fitness values.

Enums

RunResult
The result of running an entire simulation.

StepResult
The result of running a single step.

Traits

Builder
A Builder can create new instances of an object. For this library, only Simulation objects use this Builder.

Simulation
A Simulation is an execution of a genetic algorithm.

Type Definitions

NanoSecond
Simulation run time is defined in nanoseconds.

SimResult
The result of a simulation, containing the best phenotype or an error message.